Refactor LearnPost entity and types: move type definitions to learn-post.types.ts and update imports in LearnPostBuilderService for improved organization and clarity.
